refactor: dto 관리할 condition package 생성 후 옮기기
