+++ src/main/java/com/prgrms/tenwonmoa/domain/accountbook/repository/AccountBookQueryRepository.java
@@ -31,9 +31,9 @@ import com.prgrms.tenwonmoa.domain.accountbook.dto.FindCalendarResponse;
 import com.prgrms.tenwonmoa.domain.accountbook.dto.FindDayAccountResponse;
 import com.prgrms.tenwonmoa.domain.accountbook.dto.FindMonthAccountResponse;
 import com.prgrms.tenwonmoa.domain.accountbook.dto.FindSumResponse;
-import com.prgrms.tenwonmoa.domain.accountbook.dto.MonthCondition;
 import com.prgrms.tenwonmoa.domain.accountbook.dto.MonthDetail;
-import com.prgrms.tenwonmoa.domain.accountbook.dto.YearMonthCondition;
+import com.prgrms.tenwonmoa.domain.accountbook.dto.condition.MonthCondition;
+import com.prgrms.tenwonmoa.domain.accountbook.dto.condition.YearMonthCondition;
 import com.prgrms.tenwonmoa.domain.category.CategoryType;
 import com.prgrms.tenwonmoa.domain.common.page.PageCustomImpl;
 import com.prgrms.tenwonmoa.domain.common.page.PageCustomRequest;
