+++ src/main/java/com/prgrms/tenwonmoa/domain/accountbook/service/AccountBookQueryService.java
@@ -9,8 +9,8 @@ import com.prgrms.tenwonmoa.domain.accountbook.dto.FindCalendarResponse;
 import com.prgrms.tenwonmoa.domain.accountbook.dto.FindDayAccountResponse;
 import com.prgrms.tenwonmoa.domain.accountbook.dto.FindMonthAccountResponse;
 import com.prgrms.tenwonmoa.domain.accountbook.dto.FindSumResponse;
-import com.prgrms.tenwonmoa.domain.accountbook.dto.MonthCondition;
-import com.prgrms.tenwonmoa.domain.accountbook.dto.YearMonthCondition;
+import com.prgrms.tenwonmoa.domain.accountbook.dto.condition.MonthCondition;
+import com.prgrms.tenwonmoa.domain.accountbook.dto.condition.YearMonthCondition;
 import com.prgrms.tenwonmoa.domain.accountbook.repository.AccountBookQueryRepository;
 import com.prgrms.tenwonmoa.domain.common.page.PageCustomRequest;
 import com.prgrms.tenwonmoa.domain.common.page.PageResponse;
