+++ src/test/java/com/prgrms/tenwonmoa/domain/accountbook/repository/AccountBookQueryRepositoryTest.java
@@ -23,9 +23,9 @@ import com.prgrms.tenwonmoa.domain.accountbook.dto.FindCalendarResponse;
 import com.prgrms.tenwonmoa.domain.accountbook.dto.FindDayAccountResponse;
 import com.prgrms.tenwonmoa.domain.accountbook.dto.FindMonthAccountResponse;
 import com.prgrms.tenwonmoa.domain.accountbook.dto.FindSumResponse;
-import com.prgrms.tenwonmoa.domain.accountbook.dto.MonthCondition;
 import com.prgrms.tenwonmoa.domain.accountbook.dto.MonthDetail;
-import com.prgrms.tenwonmoa.domain.accountbook.dto.YearMonthCondition;
+import com.prgrms.tenwonmoa.domain.accountbook.dto.condition.MonthCondition;
+import com.prgrms.tenwonmoa.domain.accountbook.dto.condition.YearMonthCondition;
 import com.prgrms.tenwonmoa.domain.category.Category;
 import com.prgrms.tenwonmoa.domain.category.UserCategory;
 import com.prgrms.tenwonmoa.domain.common.page.PageCustomRequest;
